safecross-demo/backend/requirements.txt
fastapi
uvicorn[standard]
aiosqlite
pydantic
httpx
google-generativeai
python-dotenv
Pillow
